table updates
