+++ projects/test-app/src/app/default-example/controls/controls.component.scss
@@ -1,33 +1,33 @@
 :host {
-   display: block;
-   width: 100%;
-   .form-control {
+  display: block;
+  width: 100%;
+  .form-control {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: flex-start;
+    gap: 0.5em;
+    label {
+      font-weight: bold;
+      font-size: 0.9em;
+    }
+    input,
+    textarea {
       width: 100%;
-      display: flex;
-      flex-direction: column;
-      align-items: flex-start;
-      justify-content: flex-start;
-      gap: 0.5em;
+    }
+  }
+
+  &.invalid {
+    .form-control {
       label {
-         font-weight: bold;
-         font-size: 0.9em;
+        color: red;
       }
       input,
       textarea {
-         width: 100%;
+        border-color: red;
+        outline: none;
       }
-   }
-
-   &.invalid {
-      .form-control {
-         label {
-            color: red;
-         }
-         input,
-         textarea {
-            border-color: red;
-            outline: none;
-         }
-      }
-   }
+    }
+  }
 }
