+++ projects/test-app/src/app/default-example/default-example.component.scss
@@ -1,33 +1,24 @@
 :host {
-   display: block;
-   height: 100%;
-   background-color: gray;
-   padding: 2px;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+  background-color: gray;
+  padding: 2px;
+  gap: 2px;
 }
-.content {
-   display: flex;
-   flex-direction: column;
-   justify-content: flex-start;
-   align-items: flex-start;
-   height: 100%;
-   gap: 0.25em;
-   .overall-info {
-      width: 100%;
-   }
-   .form-controls {
-      width: 100%;
-      display: flex;
-      flex-wrap: wrap;
-      gap: 0.25em;
-   }
 
-   form-info,
-   name-control,
-   email-control,
-   message-control {
-      padding: 1em;
-      // border-bottom: 3px solid rgb(44, 206, 255);
-      border-radius: 0.5em;
-      background-color: white;
-   }
+.form-controls {
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5em;
+}
+
+form-info,
+name-control,
+email-control,
+message-control {
+  padding: 1em;
+  background-color: white;
 }
